Primeira refatoração das pastas (módulos) do sistema.
