+++ FarmaBot/FarmaBot.csproj
@@ -52,35 +52,35 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Bot\Callbacks\AdicionarCarrinhoCallback.cs" />
-    <Compile Include="Bot\Callbacks\BotBaseCallback.cs" />
-    <Compile Include="Bot\Callbacks\BotCallbackFactory.cs" />
-    <Compile Include="Bot\Callbacks\BotCallbackType.cs" />
-    <Compile Include="Bot\Callbacks\FinalizarPedidoCallback.cs" />
-    <Compile Include="Bot\Commands\BotCommandFactory.cs" />
-    <Compile Include="Bot\Commands\FarmaBotCommandBase.cs" />
-    <Compile Include="Bot\Commands\InvalidBotCommand.cs" />
-    <Compile Include="Bot\Commands\PedidoBotCommand.cs" />
-    <Compile Include="Bot\Commands\StartBotCommand.cs" />
-    <Compile Include="Bot\Commands\DiagnoseBotCommand.cs" />
-    <Compile Include="Bot\FarmaBot.cs" />
-    <Compile Include="Data\LiteDbDatabase.cs" />
-    <Compile Include="Data\MemoryDatabase.cs" />
-    <Compile Include="Models\CarrinhoCompras.cs" />
-    <Compile Include="Models\Endereco.cs" />
-    <Compile Include="Repository\IBotRepository.cs" />
-    <Compile Include="Models\Medicamento.cs" />
-    <Compile Include="Models\Pedido.cs" />
-    <Compile Include="Models\Sintoma.cs" />
+    <Compile Include="UI\Callbacks\AdicionarCarrinhoCallback.cs" />
+    <Compile Include="UI\Callbacks\BotBaseCallback.cs" />
+    <Compile Include="UI\Callbacks\BotCallbackFactory.cs" />
+    <Compile Include="UI\Callbacks\BotCallbackType.cs" />
+    <Compile Include="UI\Callbacks\FinalizarPedidoCallback.cs" />
+    <Compile Include="UI\Commands\BotCommandFactory.cs" />
+    <Compile Include="UI\Commands\FarmaBotCommandBase.cs" />
+    <Compile Include="UI\Commands\InvalidBotCommand.cs" />
+    <Compile Include="UI\Commands\PedidoBotCommand.cs" />
+    <Compile Include="UI\Commands\StartBotCommand.cs" />
+    <Compile Include="UI\Commands\DiagnoseBotCommand.cs" />
+    <Compile Include="UI\FarmaBot.cs" />
+    <Compile Include="Infra\Data\LiteDbDatabase.cs" />
+    <Compile Include="Infra\Data\MemoryDatabase.cs" />
+    <Compile Include="Model\CarrinhoCompras.cs" />
+    <Compile Include="Model\Endereco.cs" />
+    <Compile Include="Infra\Repository\IBotRepository.cs" />
+    <Compile Include="Model\Medicamento.cs" />
+    <Compile Include="Model\Pedido.cs" />
+    <Compile Include="Model\Sintoma.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Repository\LiteDBBotRepository.cs" />
-    <Compile Include="Repository\MemoryRepository.cs" />
-    <Compile Include="Services\BotService.cs" />
-    <Compile Include="Services\IBotService.cs" />
-    <Compile Include="Models\Sessao.cs" />
-    <Compile Include="Services\SessionManager.cs" />
-    <Compile Include="Utils\StringExtensions.cs" />
+    <Compile Include="Infra\Repository\LiteDBBotRepository.cs" />
+    <Compile Include="Infra\Repository\MemoryRepository.cs" />
+    <Compile Include="App\BotService.cs" />
+    <Compile Include="App\IBotService.cs" />
+    <Compile Include="Model\Sessao.cs" />
+    <Compile Include="App\SessionManager.cs" />
+    <Compile Include="Infra\StringExtensions.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
