+++ FarmaBot/Infra/Repository/IBotRepository.cs
@@ -1,0 +1,11 @@
+﻿using FarmaBot.Model;
+using System.Collections.Generic;
+
+namespace FarmaBot.Infra.Repository
+{
+    public interface IBotRepository
+    {
+        List<Medicamento> BuscarMedicamentos(Sintoma sintoma);
+        int RealizarPedido(Pedido pedido);
+    }
+}
