+++ FarmaBot/Model/Endereco.cs
@@ -1,0 +1,8 @@
+﻿namespace FarmaBot.Model
+{
+    public class Endereco
+    {
+        public float Latitude { get; set; }
+        public float Longitude { get; set; }
+    }
+}
