+++ FarmaBot/Model/Sessao.cs
@@ -1,0 +1,18 @@
+﻿namespace FarmaBot.Model
+{
+    public class Sessao
+    {
+        public Sessao()
+        {
+            Carrinho = new CarrinhoCompras();
+        }
+
+        public void AlterarLocalizacao(Endereco endereco)
+        {
+            Endereco = endereco;
+        }
+
+        public CarrinhoCompras Carrinho { get; private set; }
+        public Endereco Endereco { get; private set; }
+    }
+}
