+++ FarmaBot/Model/Sintoma.cs
@@ -1,0 +1,21 @@
+﻿using System.Collections.Generic;
+
+namespace FarmaBot.Model
+{
+    public class Sintoma
+    {
+        public int Id { get; set; }
+        public string Descricao { get; set; }
+        public List<Medicamento> Medicamentos { get; set; } = new List<Medicamento>();
+
+        public List<string> AsList()
+        {
+            var result = new List<string>();
+
+            if (string.IsNullOrEmpty(Descricao)) return result;
+
+            result.AddRange(Descricao.Split(','));
+            return result;
+        }
+    }
+}
