--- FarmaBot/Models/Endereco.cs
@@ -1,8 +1,0 @@
-﻿namespace FarmaBot.Models
-{
-    public class Endereco
-    {
-        public float Latitude { get; set; }
-        public float Longitude { get; set; }
-    }
-}
