+++ FarmaBot/UI/Callbacks/BotBaseCallback.cs
@@ -1,0 +1,13 @@
+﻿using Telegram.Bot;
+using Telegram.Bot.Types;
+
+namespace FarmaBot.UI.Callbacks
+{
+    public abstract class BotBaseCallback
+    {
+        public abstract new BotCallbackType GetType();
+        public abstract void Execute(CallbackQuery callbackQuery);
+
+        public TelegramBotClient Bot { get; set; }
+    }    
+}
