--- FarmaBot/Utils/StringExtensions.cs
@@ -1,16 +1,0 @@
-﻿using System.Globalization;
-using System.Linq;
-using System.Text;
-
-namespace FarmaBot.Utils
-{
-    public static class StringExtensions
-    {
-        public static string Raw(this string values)
-        {
-            return new string(values.Trim().ToLowerInvariant().Normalize(NormalizationForm.FormD)
-                .Where(ch => char.GetUnicodeCategory(ch) != UnicodeCategory.NonSpacingMark)
-                .ToArray());
-        }
-    }
-}
